doc add undo break
Without <c-g>u for imaps (insert-mode mappings), contiguous jumps in
insert mode to the next or prev placeholders will all be in one undo
change. Thus, by adding <c-g>u in insert mode just before selecting the
next or prev placeholder, the dilemma is fixed. Of course, the user is
still allowed to create whatever mappings she desires; the <plug>
mappings still only select the placeholders. This undo break in the docs
is just for clarity.